% Auto-generated: do not edit by hand
\name{forestThreejs}

\alias{forestThreejs}

\title{ForestThreejs component}

\description{

}

\usage{
forestThreejs(id=NULL, spacing=NULL, stats=NULL, totalX=NULL, totalZ=NULL)
}

\arguments{
\item{id}{Character. }

\item{spacing}{Numeric. }

\item{stats}{Logical. }

\item{totalX}{Numeric. }

\item{totalZ}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

